Fixing double convertions from rational numbers if result value in range (0, 1). Adding tests for that cases.

+++ Digimarc.NTiff/Types/SRational.cs
@@ -34,7 +34,7 @@ namespace Digimarc.NTiff.Types
         public readonly int Denominator;
         public double ToDouble()
         {
-            return Numerator / Denominator;
+            return (double)Numerator / Denominator;
         }
 
         public override string ToString()
